modularized and tested code, removed comments
