+++ sources/rlcs/message_types.py
@@ -1,0 +1,10 @@
+msg_index = [
+    "rlcs_main_batt_mv",
+    "rlcs_actuator_batt_mv",
+    "healthy_actuators",
+    "ignition_primary_ma",
+    "ignition_primary_ma",
+    "fill_valve_state",
+    "vent_valve_state",
+    "injector_valve_state"
+]
